+++ test/flutter_quill_delta_easy_parser_test.dart
@@ -19,27 +19,22 @@ void main() {
         lines: [
           Line(data: '\n'),
         ],
-        type: ParagraphType.block,
+        type: ParagraphType.lineBreak,
       ),
     ]);
 
     final Document? parsedDocument = RichTextParser().parseDelta(delta);
-    expect(
-        parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
+    expect(parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
 
     for (int i = 0; i < expectedDocument.paragraphs.length; i++) {
-      expect(parsedDocument?.paragraphs[i].lines.length,
-          expectedDocument.paragraphs[i].lines.length);
+      expect(parsedDocument?.paragraphs[i].lines.length, expectedDocument.paragraphs[i].lines.length);
       for (int j = 0; j < expectedDocument.paragraphs[i].lines.length; j++) {
-        expect(parsedDocument?.paragraphs[i].lines[j].data,
-            expectedDocument.paragraphs[i].lines[j].data);
-        expect(parsedDocument?.paragraphs[i].lines[j].attributes,
-            expectedDocument.paragraphs[i].lines[j].attributes);
+        expect(parsedDocument?.paragraphs[i].lines[j].data, expectedDocument.paragraphs[i].lines[j].data);
+        expect(
+            parsedDocument?.paragraphs[i].lines[j].attributes, expectedDocument.paragraphs[i].lines[j].attributes);
       }
-      expect(parsedDocument?.paragraphs[i].blockAttributes,
-          expectedDocument.paragraphs[i].blockAttributes);
-      expect(parsedDocument?.paragraphs[i].type,
-          expectedDocument.paragraphs[i].type);
+      expect(parsedDocument?.paragraphs[i].blockAttributes, expectedDocument.paragraphs[i].blockAttributes);
+      expect(parsedDocument?.paragraphs[i].type, expectedDocument.paragraphs[i].type);
     }
   });
 
@@ -61,28 +56,57 @@ void main() {
         lines: [
           Line(data: '\n'),
         ],
-        type: ParagraphType.block,
+        type: ParagraphType.lineBreak,
       ),
     ]);
 
     final Document? parsedDocument = RichTextParser().parseDelta(delta);
-    expect(
-        parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
+    expect(parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
 
     for (int i = 0; i < expectedDocument.paragraphs.length; i++) {
-      expect(parsedDocument?.paragraphs[i].lines.length,
-          expectedDocument.paragraphs[i].lines.length);
+      expect(parsedDocument?.paragraphs[i].lines.length, expectedDocument.paragraphs[i].lines.length);
       for (int j = 0; j < expectedDocument.paragraphs[i].lines.length; j++) {
-        expect(parsedDocument?.paragraphs[i].lines[j].data,
-            expectedDocument.paragraphs[i].lines[j].data);
-        expect(parsedDocument?.paragraphs[i].lines[j].attributes,
-            expectedDocument.paragraphs[i].lines[j].attributes);
+        expect(parsedDocument?.paragraphs[i].lines[j].data, expectedDocument.paragraphs[i].lines[j].data);
+        expect(
+            parsedDocument?.paragraphs[i].lines[j].attributes, expectedDocument.paragraphs[i].lines[j].attributes);
       }
-      expect(parsedDocument?.paragraphs[i].blockAttributes,
-          expectedDocument.paragraphs[i].blockAttributes);
-      expect(parsedDocument?.paragraphs[i].type,
-          expectedDocument.paragraphs[i].type);
+      expect(parsedDocument?.paragraphs[i].blockAttributes, expectedDocument.paragraphs[i].blockAttributes);
+      expect(parsedDocument?.paragraphs[i].type, expectedDocument.paragraphs[i].type);
     }
+  });
+
+  test('should merge similar operations that contains same attributes (even if both does not contains them)', () {
+    final Delta delta = Delta.fromOperations([
+      Operation.insert('This is an interesting example', {'bold': true}),
+      Operation.insert(' about how the easy parser can work ', {'bold': true}),
+      Operation.insert('but, sometimes, it could get a unexpected behavior,'),
+      Operation.insert(' so... we make some test to avoid that'),
+      Operation.insert('\n'),
+    ]);
+
+    final Document expectedDocument = Document(paragraphs: [
+      Paragraph(
+        lines: [
+          Line(
+              data: "This is an interesting example about how the easy parser can work ",
+              attributes: {'bold': true}),
+          Line(
+            data: "but, sometimes, it could get a unexpected behavior, so... we make some test to avoid that",
+            attributes: {'bold': true},
+          ),
+        ],
+        type: ParagraphType.inline,
+      ),
+      Paragraph(
+        lines: [
+          Line(data: '\n'),
+        ],
+        type: ParagraphType.lineBreak,
+      ),
+    ]);
+
+    final Document? parsedDocument = RichTextParser().parseDelta(delta);
+    expect(parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
   });
 
   test('Should convert aligned header to paragraph block', () {
@@ -103,27 +127,22 @@ void main() {
         lines: [
           Line(data: '\n'),
         ],
-        type: ParagraphType.block,
+        type: ParagraphType.lineBreak,
       ),
     ]);
 
     final Document? parsedDocument = RichTextParser().parseDelta(delta);
-    expect(
-        parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
+    expect(parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
 
     for (int i = 0; i < expectedDocument.paragraphs.length; i++) {
-      expect(parsedDocument?.paragraphs[i].lines.length,
-          expectedDocument.paragraphs[i].lines.length);
+      expect(parsedDocument?.paragraphs[i].lines.length, expectedDocument.paragraphs[i].lines.length);
       for (int j = 0; j < expectedDocument.paragraphs[i].lines.length; j++) {
-        expect(parsedDocument?.paragraphs[i].lines[j].data,
-            expectedDocument.paragraphs[i].lines[j].data);
-        expect(parsedDocument?.paragraphs[i].lines[j].attributes,
-            expectedDocument.paragraphs[i].lines[j].attributes);
+        expect(parsedDocument?.paragraphs[i].lines[j].data, expectedDocument.paragraphs[i].lines[j].data);
+        expect(
+            parsedDocument?.paragraphs[i].lines[j].attributes, expectedDocument.paragraphs[i].lines[j].attributes);
       }
-      expect(parsedDocument?.paragraphs[i].blockAttributes,
-          expectedDocument.paragraphs[i].blockAttributes);
-      expect(parsedDocument?.paragraphs[i].type,
-          expectedDocument.paragraphs[i].type);
+      expect(parsedDocument?.paragraphs[i].blockAttributes, expectedDocument.paragraphs[i].blockAttributes);
+      expect(parsedDocument?.paragraphs[i].type, expectedDocument.paragraphs[i].type);
     }
   });
 
@@ -159,10 +178,7 @@ void main() {
         blockAttributes: {"header": 1},
         type: ParagraphType.block,
       ),
-      Paragraph(
-          lines: [Line(data: '\n')],
-          blockAttributes: {"header": 1},
-          type: ParagraphType.block),
+      Paragraph(lines: [Line(data: '\n')], blockAttributes: {"header": 1}, type: ParagraphType.lineBreak),
       Paragraph(
         lines: [
           Line(data: 'This is a list item'),
@@ -183,7 +199,7 @@ void main() {
         ],
         type: ParagraphType.inline,
       ),
-      Paragraph(lines: [Line(data: '\n')], type: ParagraphType.block),
+      Paragraph(lines: [Line(data: '\n')], type: ParagraphType.lineBreak),
       Paragraph(
         lines: [
           Line(data: 'This is a '),
@@ -194,22 +210,17 @@ void main() {
       ),
     ]);
     final Document? parsedDocument = RichTextParser().parseDelta(delta);
-    expect(
-        parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
+    expect(parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
 
     for (int i = 0; i < expectedDocument.paragraphs.length; i++) {
-      expect(parsedDocument?.paragraphs[i].lines.length,
-          expectedDocument.paragraphs[i].lines.length);
+      expect(parsedDocument?.paragraphs[i].lines.length, expectedDocument.paragraphs[i].lines.length);
       for (int j = 0; j < expectedDocument.paragraphs[i].lines.length; j++) {
-        expect(parsedDocument?.paragraphs[i].lines[j].data,
-            expectedDocument.paragraphs[i].lines[j].data);
-        expect(parsedDocument?.paragraphs[i].lines[j].attributes,
-            expectedDocument.paragraphs[i].lines[j].attributes);
+        expect(parsedDocument?.paragraphs[i].lines[j].data, expectedDocument.paragraphs[i].lines[j].data);
+        expect(
+            parsedDocument?.paragraphs[i].lines[j].attributes, expectedDocument.paragraphs[i].lines[j].attributes);
       }
-      expect(parsedDocument?.paragraphs[i].blockAttributes,
-          expectedDocument.paragraphs[i].blockAttributes);
-      expect(parsedDocument?.paragraphs[i].type,
-          expectedDocument.paragraphs[i].type);
+      expect(parsedDocument?.paragraphs[i].blockAttributes, expectedDocument.paragraphs[i].blockAttributes);
+      expect(parsedDocument?.paragraphs[i].type, expectedDocument.paragraphs[i].type);
     }
   });
 
@@ -225,26 +236,21 @@ void main() {
     final Delta deltaWithNewlines = Delta()..insert('\n\n\n\n');
 
     final Document expectedDocument = Document(paragraphs: [
-      Paragraph(lines: [Line(data: '\n')], type: ParagraphType.block),
-      Paragraph(lines: [Line(data: '\n')], type: ParagraphType.block),
-      Paragraph(lines: [Line(data: '\n')], type: ParagraphType.block),
-      Paragraph(lines: [Line(data: '\n')], type: ParagraphType.block),
+      Paragraph(lines: [Line(data: '\n')], type: ParagraphType.lineBreak),
+      Paragraph(lines: [Line(data: '\n')], type: ParagraphType.lineBreak),
+      Paragraph(lines: [Line(data: '\n')], type: ParagraphType.lineBreak),
+      Paragraph(lines: [Line(data: '\n')], type: ParagraphType.lineBreak),
     ]);
 
-    final Document? parsedDocument =
-        RichTextParser().parseDelta(deltaWithNewlines);
-    expect(
-        parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
+    final Document? parsedDocument = RichTextParser().parseDelta(deltaWithNewlines);
+    expect(parsedDocument?.paragraphs.length, expectedDocument.paragraphs.length);
 
     for (int i = 0; i < expectedDocument.paragraphs.length; i++) {
-      expect(parsedDocument?.paragraphs[i].lines.length,
-          expectedDocument.paragraphs[i].lines.length);
+      expect(parsedDocument?.paragraphs[i].lines.length, expectedDocument.paragraphs[i].lines.length);
       for (int j = 0; j < expectedDocument.paragraphs[i].lines.length; j++) {
-        expect(parsedDocument?.paragraphs[i].lines[j].data,
-            expectedDocument.paragraphs[i].lines[j].data);
+        expect(parsedDocument?.paragraphs[i].lines[j].data, expectedDocument.paragraphs[i].lines[j].data);
       }
-      expect(parsedDocument?.paragraphs[i].type,
-          expectedDocument.paragraphs[i].type);
+      expect(parsedDocument?.paragraphs[i].type, expectedDocument.paragraphs[i].type);
     }
   });
 }