+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.5
+
+* Fix: sometimes, the `EmbedObject`s can be merged into a inline `Paragraph`.
+* Chore: now the parser divides every `Embed`, `Block`, and `NewLine` if its different types.
+* Chore: added a new type of `Paragraph` called `ParagraphType.lineBreak`.
+* Chore: deprecate `ensureCorrectFormat()` from `Document` class since it already don't do nothing. 
+* Chore: removed `setupInfo` param from `Document` since never was used.
+* Chore: removed some exports that shouldn't be part of the public API;
+
 ## 1.0.4
 
 * Fix: duplicated new lines